requirements.txt
# Function dependencies, for example:
# package>=version
requests==2.27.1
PyYAML==6.0
